requirements.txt
gphoto2~=2.3.4
PyQt6~=6.4.2
Pillow~=9.4.0